requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.1.0
matplotlib>=3.6.0
seaborn>=0.12.0
python-dotenv>=0.21.0
openpyxl>=3.0.10
xlrd>=2.0.1
# Web server & data
flask>=3.0.0
flask-cors>=4.0.0
yfinance>=0.2.36
requests>=2.31.0
